Use payment.hash as a unique ID for payment rows
Currently payment.time is used as a unique ID for payment table rows. This could be an issue if multiple payments has exactly same timestamps as only 1 payment out of all the payments with same timestamps will be shown. Using payment.hash as table row ID ensures that all the rows has a unique ID and would prevent issues that come with multiple payments with same timestamps.